SMALLINT instead of TINYINT

+++ springs-persistence-orm-users/src/main/java/com/github/justincranford/springs/persistenceorm/users/persona/EmailAddressOrm.java
@@ -28,7 +28,7 @@ import lombok.experimental.Accessors;
 @AllArgsConstructor
 @Accessors(fluent=true)
 public class EmailAddressOrm {
-	@Column(nullable=false,columnDefinition="TINYINT")
+	@Column(nullable=false,columnDefinition="SMALLINT")
 	private int rank;
 
 	@Embedded
